Split reducers and actions for better readability
