+++ src/application/reducers/mainReducer.tsx
@@ -1,0 +1,17 @@
+import {combineReducers} from 'redux'
+
+import {appData} from './appData'
+import {currentDatabase} from './currentDatabase'
+import {currentFhirAttribute} from './currentFhirAttribute'
+import {currentFhirResource} from './currentFhirResource'
+import {mapping} from './mapping'
+import {nameLists} from './nameLists'
+
+export const mainReducer = combineReducers({
+    currentDatabase,
+    currentFhirResource,
+    currentFhirAttribute,
+    mapping,
+    appData,
+    nameLists,
+})
